use std::alloc::dealloc to deallocate memory
Copy the nomicon's approach to deallocating from a Vec. Results in less
code, and it's a little easier to understand.

One commenter suggested that we should use a slice of MaybeUninit<T> for
the contents, which is a nice design. But since this code is intended
for educational purposes, I decided to keep the raw pointer as it's a
little easier to understand.